% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WithDate.R
\name{WithDateHTML}
\alias{WithDateHTML}
\title{Adds the system date to the RMD file name.}
\usage{
WithDateHTML(input, ...)
}
\arguments{
\item{input}{comes from knit (Nothinhg to do)}

\item{...}{comes from knit (Nothinhg to do)}
}
\value{

}
\description{
The knit files are saved in knit2file folder.
}
\details{
Add to yaml Header from RMD Report:
knit: Knit2File::WithDateHTML
}
\author{
quito96
}
