% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WithDate.R
\name{WithDateHTML}
\alias{WithDateHTML}
\title{Add System Date to knit File Name}
\usage{
WithDateHTML(input, ...)
}
\arguments{
\item{input}{comes from knit (Nothinhg to do)}

\item{...}{comes from knit (Nothinhg to do)}
}
\value{

}
\description{
Add System Date to knit File Name
}
\note{
Add to yaml Header from RMD Report:
knit: Knit2File::WithDateHTML
}
\author{
quito96
}
