% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WithDate.R
\name{WithDateHTML}
\alias{WithDateHTML}
\title{Custom Knit function for RStudio
Add System Date to knit File Name}
\usage{
WithDateHTML(input, ...)
}
\arguments{
\item{input}{only knit}

\item{...}{only knit}
}
\value{

}
\description{
Custom Knit function for RStudio
Add System Date to knit File Name
}
